+++ packages/graphql/tests/integration/directives/vector/vector-query.int.test.ts
@@ -31,6 +31,7 @@ describe("@vector directive - Query", () => {
 
     let driver: Driver;
     let MULTIDB_SUPPORT = true;
+    let VECTOR_SUPPORT = true;
     let neoSchema: Neo4jGraphQL;
 
     let Movie: UniqueType;
@@ -48,8 +49,15 @@ describe("@vector directive - Query", () => {
     };
 
     beforeAll(async () => {
-        const databaseName = generate({ readable: true, charset: "alphabetic" });
+        const dbInfo = await testHelper.getDatabaseInfo();
+        // No vector support, so we skip tests
+        if (!dbInfo.gte("5.15")) {
+            VECTOR_SUPPORT = false;
+            await testHelper.close();
+            return;
+        }
 
+        const databaseName = generate({ readable: true, charset: "alphabetic" });
         try {
             await testHelper.createDatabase(databaseName);
             driver = await testHelper.getDriver();
@@ -118,13 +126,19 @@ describe("@vector directive - Query", () => {
     });
 
     afterAll(async () => {
-        if (MULTIDB_SUPPORT) {
+        if (MULTIDB_SUPPORT && VECTOR_SUPPORT) {
             await testHelper.dropDatabase();
             await testHelper.close();
         }
     });
 
     test("Retrieve nodes ordered by score DESC by default", async () => {
+        // Skip if vector not supported
+        if (!VECTOR_SUPPORT) {
+            console.log("VECTOR SUPPORT NOT AVAILABLE - SKIPPING");
+            return;
+        }
+
         // Skip if multi-db not supported
         if (!MULTIDB_SUPPORT) {
             console.log("MULTIDB_SUPPORT NOT AVAILABLE - SKIPPING");
@@ -171,6 +185,12 @@ describe("@vector directive - Query", () => {
     });
 
     test("Retrieve nodes with relationships", async () => {
+        // Skip if vector not supported
+        if (!VECTOR_SUPPORT) {
+            console.log("VECTOR SUPPORT NOT AVAILABLE - SKIPPING");
+            return;
+        }
+
         // Skip if multi-db not supported
         if (!MULTIDB_SUPPORT) {
             console.log("MULTIDB_SUPPORT NOT AVAILABLE - SKIPPING");
